Add MVP for ChaptersViewController
